Pattern Creator: Move settings sidebar to a "pre-submission" panel
The pattern settings are updated before submission/update, so they only need to show up in a pre-submission panel. This moves the settings into the "actions" section of the InterfaceSkeleton, has it show up on the first "Submit" press, and then once the panel is open the real save button appears. This also contains some fixes for style inconsistencies.