Pattern CPT API: Add a field to pass through the raw block content
The raw content is needed because it contains the source code for blocks (the comment delimiters). The rendered content is considered a "classic block", since it lacks these. The `edit` context would return both raw and rendered, but it requires more permissions and potentially exposes more content than we need. Instead, we can create a new field which simply passes through the raw content.