Theme: Add live pattern preview to single pattern page (#46)
Fixes #43 — Create a single pattern template & preview component. This adds an iframe to "containerize" the preview's CSS, and let us resize it like an actual browser would. I've pulled in Twenty Twenty-One's CSS from wp-themes.com, but left a note that we could swap this with a local file or from another theme as needed.