Directory: Add a system for flagging patterns (#77)
This introduces the wporg-pattern-flag post type, the wporg-pattern-flag-reason taxonomy, and the resolved post status to store and manage data when a pattern in the pattern directory is flagged for review. The post type includes a REST API endpoint, which is a slightly modified version of the standard endpoint for WP post types that allows front end users to create new flags as long as they are logged into their wp.org account. This also includes back-end UI for managing flags and taxonomy terms, namely a list table and a UI for creating new terms.

Refs #61