Add locale picker to pattern editor (#229)
This adds a locale field to the Pattern Details panel in the pattern editor (front and back end). The chosen locale is saved as a postmeta value (`wpop_locale`) on the pattern post. The locale data that populates the field comes from the same mu-plugin used elsewhere on dotorg to supply locale data, so this also updates the local dev configuration to contain the required files and database table.

Refs #223, #224