Pattern Creator: Set up creating/editing permissions for logged in wp.org users (#249)
Enables logged in wp.org users to create and edit patterns on the front end of the Pattern Directory without having to be an actual member of the patterns site. It accomplishes this by doing some last-minute filtering of a user's capabilities in some situations. As such, this also includes some capability-related unit tests.

This also fixes an issue with the pattern theme's stylesheet that was preventing it from loading correctly when trying to access the New Pattern screen while logged out.

Fixes #30