Report pattern modal: Sort reason terms by slug for custom sort order
It makes the most sense for the "Other" reason to be last in the list,
but since it was added first, and we might add other reasons later,
sorting by the ID makes it hard to put it there. Sorting by slug will
make this easier because we can customize the slug through the UI to
be things like `1-inappropriate`, `2-other` etc.