Patterns: Handle empty results and 404 pages (#280)
* Pattern Grid: Rename loading state component.

This pulls the grid-logic out of the `PatternGrid` and allows the length of the skeleton to be customized to the query request.

* Patterns: Show a notice state when Patterns has no content.

* Add 404 page

* Add the empty state to "My Patterns" and search results

* Show an empty state on favorites

This changes the initial state for `favorites` to `null`, so we can tell the difference between "not loaded" and "empty". Use this null value, in combination with the patterns query, to check if there are any favorite patterns - otherwise show the empty state.