Stats snapshots for patterns (#294)
Adds functions for gathering stats about patterns and saving them to snapshot posts, similar to what is done on the 5ftf site. A cron job is set up to run at 00:00 UTC daily to save a snapshot. Also adds a simple, work-in-progress page in WP Admin to view the current stats and some basic info about the number of snapshots that have been created.

Fixes #292