Navigation Menu: Fix spacing around menu and message (#338)
* Adjust margins on Nav Layout and Context Bar components

The margins will collapse together when both components are visible, and if not, they will still have the appropriate spacing. This also cleans up some of the responsive styling, to only override the changed values.

* Remove the fixed height from ContextBar

This was causing a double-margin on the homepage (when height is 0). Instead, since we're not doing the height animation anymore, we can hide the whole component when there is no message.

* Remove the ref and container div