Tooling: Update the version of the Push action for the i18n workflow
The commit hash we have been pinning for this action had a bug that
caused the commit message to be wrapped in curly brackets. This new
commit hash is a new version release that includes a fix for that and
one other issue (I reviewed all the changes).

See https://github.com/actions-js/push/issues/3