Patterns: Focus on the context bar after navigation (#356)
This triggers the browser to scroll back to the top of the page, moves the keyboard focus, and announces to screen reader users that the page content has changed. The ContextBar content is now always on screen, but sometimes (ex, on the homepage), it is visually hidden. There is also now a "Page X of Y." label in the ContextBar, but this is only read by screen readers.

Fixes #297.