Patterns: Switch to using meta value for pattern keywords
This deprecates the "Keywords" taxonomy by labelling it internal, as it's still used for the "Core" flag, and could be used elsewhere in the future. The "Pattern Details" panel now has a keywords section that tokenizes keywords like the frontend editor does. Then it saves the keywords as a comma-separated string. Finally, add the keywords meta field to the ES safelist.