Validation: Ensure patterns use chosen default public status
The default status can be "publish" or "pending", but it was possible to override the default status and set patterns to "publish" if you made a direct API request. This checks the post status if it's changed during an API request, and throws an error if it's moving to an invalid status.