Add a new endpoint for pattern previews (#521)
* Patterns: Add a `view` endpoint which loads only the pattern content

* Theme: Use the `view` endpoint in the preview iframes

This removes the special handling we added for the Iframe component, and these can now be plain `iframe` elements.

* Remove the inline style workarounds

Using the view endpoint, patterns are loaded with all the relevant hooks. The styles provided by the Gutenberg style engine no longer need workarounds to be inlined. This also fixes conflicts in the theme itself.

* Style preview to be vertically centered in the viewport

* Remove editorStyles workaround

* Move CSS out of the HTML itself to add a comment

* Prevent links in pattern from being clickable

* Load correct theme on wp.org network