Use `mshots` service to generate pattern screenshots (#528)
* Use `mshots` service to generate pattern screenshots

* Remove the `Disabled` wrapper

Since this isn't an iframe anymore, there isn't anything possibly-interactive we need to disable.

* Add the modified date to the URL

This will break the image cache if the pattern is updated

* Update the spinner size

* Move the size values to Screenshot

Set both image width and viewport size, so the screenshot is taken at 1200px wide, but then resized down to a 600px image before being sent back to the browser.

* Add global "thumnailVersion" cachebuster for mshots image cache

* Fetch the screenshot from the production site if this is a local env

This fixes most of the "403" errors when using the exported data, since the same patterns exist in production. The exception is any locally-created patterns, which now show a 404 page. Still, this is an improvement over errors everywhere locally.