Notifications: Only send email notifications on original patterns.
If a pattern is updated, it needs to go back through moderation. The pattern status also syncs to the translated patterns, so they also get marked back to pending. When the original pattern is published or unlisted, that also cascades down to the translated patterns. This triggers a batch of emails to the pattern author, one for each language of pattern.

The only pattern that should trigger notifications is the original, so this checks the locale before triggering any emails.

Fixes #560.