Pattern Grid: Use mshots for published patterns only
Draft & pending patterns are not accessible by the mshots service, so it screencaps the 404 page (currently a white screen). This brings back the iframe previewer for non-published patterns, leaving the performance benefit of images for the general archive. The draft/pending patterns are visible to the current user via the iframe because it uses the current user's permissions.

Fixes #543